Fix signed-integer-overflow UB in big-endian 32-bit reads across the format parsers (#293)
* Fix signed-integer-overflow UB in big-endian 32-bit reads (parsers)

The format parsers assemble 32-bit big-endian values a byte at a time, e.g.
`tmp_val = *midi_data++ << 24;`. The uint8_t operand promotes to int, so when
the top byte's high bit is set, `value << 24` shifts into/past the int sign
bit -- undefined behavior (UBSan: "left shift of 255 by 24 places cannot be
represented in type 'int'"). This fires on the normal MIDI/XMI/MUS/HMP/HMI
loading paths and when reading GUS patch headers; found with AFL++ + UBSan.

Cast the shifted byte to uint32_t so the shift is well-defined unsigned
arithmetic. The computed value is unchanged on two's-complement targets; only
the `<< 24` term can overflow (`<< 16` / `<< 8` on a byte stay within int), so
those are left as-is. 25 sites across f_midi, f_xmidi, f_hmi, f_hmp, gus_pat,
and mus2mid.


* mus2mid: cast all READ_INT32 operands to uint32_t (consistency)

Per review (@sezero): match hmi2mid.c's macro style by casting every byte
rather than only the <<24 term. No behavioural change.


---------

+++ src/f_hmp.c
@@ -111,7 +111,7 @@ _WM_ParseNewHmp(const uint8_t *hmp_data, uint32_t hmp_size) {
     hmp_file_length = *hmp_data++;
     hmp_file_length += (*hmp_data++ << 8);
     hmp_file_length += (*hmp_data++ << 16);
-    hmp_file_length += (*hmp_data++ << 24);
+    hmp_file_length += ((uint32_t)*hmp_data++ << 24);
     hmp_size -= 4;
 
     WMIDI_UNUSED(hmp_file_length);
@@ -123,7 +123,7 @@ _WM_ParseNewHmp(const uint8_t *hmp_data, uint32_t hmp_size) {
     hmp_chunks = *hmp_data++;
     hmp_chunks += (*hmp_data++ << 8);
     hmp_chunks += (*hmp_data++ << 16);
-    hmp_chunks += (*hmp_data++ << 24);
+    hmp_chunks += ((uint32_t)*hmp_data++ << 24);
     hmp_size -= 4;
 
     if (!hmp_chunks) {
@@ -135,7 +135,7 @@ _WM_ParseNewHmp(const uint8_t *hmp_data, uint32_t hmp_size) {
     hmp_unknown = *hmp_data++;
     hmp_unknown += (*hmp_data++ << 8);
     hmp_unknown += (*hmp_data++ << 16);
-    hmp_unknown += (*hmp_data++ << 24);
+    hmp_unknown += ((uint32_t)*hmp_data++ << 24);
     hmp_size -= 4;
 
     WMIDI_UNUSED(hmp_unknown);
@@ -147,7 +147,7 @@ _WM_ParseNewHmp(const uint8_t *hmp_data, uint32_t hmp_size) {
     hmp_bpm = *hmp_data++;
     hmp_bpm += (*hmp_data++ << 8);
     hmp_bpm += (*hmp_data++ << 16);
-    hmp_bpm += (*hmp_data++ << 24);
+    hmp_bpm += ((uint32_t)*hmp_data++ << 24);
     hmp_size -= 4;
 
     if (!hmp_bpm) {
@@ -171,7 +171,7 @@ _WM_ParseNewHmp(const uint8_t *hmp_data, uint32_t hmp_size) {
     hmp_song_time = *hmp_data++;
     hmp_song_time += (*hmp_data++ << 8);
     hmp_song_time += (*hmp_data++ << 16);
-    hmp_song_time += (*hmp_data++ << 24);
+    hmp_song_time += ((uint32_t)*hmp_data++ << 24);
     hmp_size -= 4;
 
     /* DEBUG */
@@ -211,7 +211,7 @@ _WM_ParseNewHmp(const uint8_t *hmp_data, uint32_t hmp_size) {
         chunk_num = *hmp_data++;
         chunk_num += (*hmp_data++ << 8);
         chunk_num += (*hmp_data++ << 16);
-        chunk_num += (*hmp_data++ << 24);
+        chunk_num += ((uint32_t)*hmp_data++ << 24);
         chunk_ofs[i] += 4;
 
         WMIDI_UNUSED(chunk_num);
@@ -219,7 +219,7 @@ _WM_ParseNewHmp(const uint8_t *hmp_data, uint32_t hmp_size) {
         chunk_length[i] = *hmp_data++;
         chunk_length[i] += (*hmp_data++ << 8);
         chunk_length[i] += (*hmp_data++ << 16);
-        chunk_length[i] += (*hmp_data++ << 24);
+        chunk_length[i] += ((uint32_t)*hmp_data++ << 24);
         chunk_ofs[i] += 4;
 
         if (chunk_length[i] > hmp_size) {
@@ -232,7 +232,7 @@ _WM_ParseNewHmp(const uint8_t *hmp_data, uint32_t hmp_size) {
         hmp_track = *hmp_data++;
         hmp_track += (*hmp_data++ << 8);
         hmp_track += (*hmp_data++ << 16);
-        hmp_track += (*hmp_data++ << 24);
+        hmp_track += ((uint32_t)*hmp_data++ << 24);
         chunk_ofs[i] += 4;
 
         WMIDI_UNUSED(hmp_track);
